Update README with Getting Started section referencing node-elm-test
This change adds a small Getting Started section to the beginning of
the README, directing users to node-elm-test.

The rationale behind this is that node-elm-test is currently
the quickest & most straightforward way to get started with Elm Test,
and yet it isn't readily discoverable from the Elm Test repo. I think
that a number of the problems and questions raised in Github Issues
and in the elm-discuss mailing list could be pre-empted by this.

We should direct folks to node-elm-test immediately so they have a
hassle-free initial experience getting Elm Test set up.