Begin total revamp of API, designed by @rtfeldman
So far, we've got a Test tree which can include various decorations,
and that gets flattened into a plan. The plan probably needs to include
the depth of the node in the original tree so the output can be indented and
failWiths from higher up are printed once and in order.

The other concern I have with the Test tree is that it seems we're making
functions both when batching and when wrapping up a list of unit tests.

I'd like to see something compilable as quickly as possible to flush out all
the wrong assumptions. To that end I'd like to get a very basic fuzzer
actually working soon.