Flesh out API enough that examples compile.
@rtfeldman - I'm not sure the Plan type is going to work. Consider that a call to
onFail buried in the thunk passed in a list to Test.unit. This won't be evaluated until
the test is run! It can't be accessed to turn it into a plan prior to running.

Therefore my current idea is to keep the tree structure and somehow map over it to get
a parallel tree, except the thunks are replaced by already-evaluated results. Alternatively,
instead of these decorator nodes we could just have Test, Group, Batch, and each of those will
have a run count, failWith, etc.

I've also removed the Seed from the tree. I think a better API than sticking seeds anywhere
is to provide one to the function that actually runs the tests. If we do manage to get tests
flattened into a list, Random.Pcg allows us to easily generate a matching list of independent
seeds, so zip over them, done.

I'm feeling pretty good about this - just needs a lot of thought, trial, and error. I don't think
we're going to hit a "it's not possible, proof attached" sort of ceiling.