Working fuzz tests.
@rtfeldman - The best data structure I could come up with is a tree
where instead of having some tags be decorators, we keep certain
information at each node. One of the driving constraints was that
assertEqual and unit/fuzz must have the same return type so that you
can use onFail with them both.

This commit also introduces Random.Pcg to give independent tests
independent seeds. For many copies of a fuzz test, it's possible to
avoid doing this and generating them sequentially, wrapped up in the
"list" combinator. Point being, I still think most of the work is
being done by the run function, not the test generating functions, so
we're on target for parallelization. Also, a single seed is provided
at the top level.

The display and results tree are fairly rudimentary, but the FuzzNode
tag ought to allow us to detect fuzz tests and avoid counting them N
times.

Still todo: shrinking (and unifying different runs of a fuzz tests).
I'm thinking of moving elm-shrink into an internal module, and then
reimplementing the parts of elm-random-extra we need for Random.Pcg -
I can also add them to the library proper if they turn out to be
useful functions.

Richard, please give this your best shot with the webworkers and
hopefully the implementation meets all the design criteria. Otherwise
we can keep tinkering...