intRange and floatRange always test endpoints 10% of the time
The purpose of a fuzzer is to pick "interesting"
values in hopes that they cause a test failure.
For a range, this means the endpoints; thankfully
fuzzer ranges are inclusive. For floats, the odds
are that the endpoints will never be tested if you
sample a range, so this makes sure we get them.
This strategy is already used by `percentage`.

If you're fuzzing over just the range, then you
don't need to produce the endpoints ten times;
once is enough. But if you're fuzzer is used to
build other fuzzers, you want some possibility of
getting pairs of interesting values.