Avoid running the test one extra time when shrinking
When inserting into the dictionary, we call getExpectation on a value
that we've already called it on. By saving the value and passing it around
we avoid calling the test again, which seems to get us a little bit of perf
and makes tests with Debug.log easier to debug.

See: https://github.com/rtfeldman/node-test-runner/issues/53#issuecomment-243655779