Add another test to confirm that manual fuzzing is working
The problem is that, in mapping (*) 2 over the fuzzer, the shrunken
values are not as granular as I would have liked. I think this is a
fundamental limitation of Rosetree fuzzing, i.e. where no inverse of the
mapped function is given and all shrunken values are (lazily)
precomputed before any mapping happens.

So, after seeing the other test pass, and a Debug.log revealed that no
shrunken value was found for 4, I changed the expectation and the fuzzer
so I'm okay with it. Honestly, finding "e" in a long string is a harder
task, and it's a good test of the fuzzer/shrinker that it can do that.

(There's a certain art to testing fuzzers in that you have to pick
claims that you know are false, but not trivially false. Claiming that
all strings contain "e" is easily disproven by the empty string, but
claiming that _no_ strings contain it forces you to find "e" on its
own.)