Bugfix: finding memoized value skips remaining tests
Fix issue where hitting a known fuzzed value would skip the rest of the
tests. Instead, do nothing for that test, but continue with the
remaining tests.

It's tempting to not count this as a test at all to ensure we hit n
distinct results, but not all fuzzers generate 100 distinct values (e.g.
Fuzz.bool).